Migrate money from Double to exact Decimal
Monetary values were stored and summed as binary Double, which drifts on
summation (0.1 + 0.2 != 0.3) and compares unreliably against targets — the
half-cent meets() epsilon in Progress.swift was an interim mitigation. Move
all currency/rate fields to base-10 Decimal so arithmetic and equality are
exact, and remove the epsilon.

Models (Values/Settings/NodeData):
- SourcedNumber.value, Debt.balance/apr/minPayment, Goal.target/saved/
  horizonYears, IRAData/HSAData.annualLimit, SavePurchaseData.target,
  CollegeData.monthlyContribution, and the five Settings money fields → Decimal.
- The Match / Increase401k percentages also move to Decimal so PercentField
  and progressOf are uniformly exact.
- emergencyFundTarget / bigEmergencyFundTarget return Decimal (months: Int).

Domain (Progress):
- Aggregate and compare in exact Decimal; drop readyTolerance/meets() for
  plain >=. ProgressInfo keeps its Double display shape (value/max convert
  via Decimal.asDouble), so GoalBar and the screens are unchanged.

App (Fields/UIKit/Editors/NodeForms):
- NumberField/PercentField parse text straight to Decimal (POSIX locale,
  never via Double) and format through Decimal-aware Currency/NumberFormat.

Wire format unchanged: a Decimal Codable-encodes as a bare JSON number
identical to the web's `number`, so on-device JSON and export/import stay
byte-compatible with src/state/schema.ts — no web/schema changes.

Tests:
- Add JSONWireFormatTests: money encodes as a JSON number (not string/
  object), web-written numbers decode to exact Decimals (no drift), the
  round-trip is byte-stable, and a Double-drifting sum stays exact.
- Update ProgressTests to assert exactness instead of tolerance; construct
  fractional test values via Decimal(string:) to avoid the float-literal
  -> Double pitfall.

+++ ios/FinanceFlowKit/Sources/FinanceFlowKit/Models/Settings.swift
@@ -2,18 +2,18 @@ import Foundation
 
 /// User-level settings that drive targets. Mirrors `Settings` in `src/state/schema.ts`.
 public struct Settings: Codable, Equatable, Sendable {
-    public var monthlyExpenses: Double?
-    public var preTaxIncome: Double?
-    public var iraAnnualLimit: Double
-    public var hsaSelfLimit: Double
-    public var hsaFamilyLimit: Double
+    public var monthlyExpenses: Decimal?
+    public var preTaxIncome: Decimal?
+    public var iraAnnualLimit: Decimal
+    public var hsaSelfLimit: Decimal
+    public var hsaFamilyLimit: Decimal
 
     public init(
-        monthlyExpenses: Double? = nil,
-        preTaxIncome: Double? = nil,
-        iraAnnualLimit: Double,
-        hsaSelfLimit: Double,
-        hsaFamilyLimit: Double
+        monthlyExpenses: Decimal? = nil,
+        preTaxIncome: Decimal? = nil,
+        iraAnnualLimit: Decimal,
+        hsaSelfLimit: Decimal,
+        hsaFamilyLimit: Decimal
     ) {
         self.monthlyExpenses = monthlyExpenses
         self.preTaxIncome = preTaxIncome
@@ -34,14 +34,14 @@ public struct Settings: Codable, Equatable, Sendable {
 
 /// Small emergency-fund target: max($1,000, one month expenses).
 /// Mirrors `emergencyFundTarget` in `src/state/schema.ts`.
-public func emergencyFundTarget(monthlyExpenses: Double?) -> Double {
+public func emergencyFundTarget(monthlyExpenses: Decimal?) -> Decimal {
     guard let monthlyExpenses, monthlyExpenses > 0 else { return 1000 }
     return max(1000, monthlyExpenses)
 }
 
 /// Big emergency-fund target: months × monthly expenses (0 if expenses unset).
 /// Mirrors `bigEmergencyFundTarget` in `src/state/schema.ts`.
-public func bigEmergencyFundTarget(months: Double, monthlyExpenses: Double?) -> Double {
+public func bigEmergencyFundTarget(months: Int, monthlyExpenses: Decimal?) -> Decimal {
     guard let monthlyExpenses, monthlyExpenses > 0 else { return 0 }
-    return months * monthlyExpenses
+    return Decimal(months) * monthlyExpenses
 }
