+++ ios/FinanceFlowKit/Sources/FinanceFlowKit/Models/Values.swift
@@ -1,20 +1,36 @@
 import Foundation
 
+// MARK: - Money representation
+//
+// Currency, rates, and other exact-decimal quantities are stored as `Decimal`
+// (base-10) rather than `Double` (base-2). `Double` can't represent values like
+// 0.1 exactly, so sums drift (0.1 + 0.2 != 0.3) and equality against a target is
+// unreliable; `Decimal` is exact for the decimal fractions money actually uses.
+//
+// Wire format: a `Decimal` Codable-encodes as a bare JSON number (e.g. `22.9`),
+// byte-identical to what the web's `number` fields produce — so the on-device
+// JSON and export/import stay compatible with `src/state/schema.ts`. See
+// `JSONWireFormatTests` for the round-trip proof.
+//
+// CAUTION: never construct these from a floating-point literal — Swift routes
+// `Decimal`'s float-literal init through `Double`, so `22.9` would store
+// 22.8999999999999986… Use integer literals (exact) or `Decimal(string:)`.
+
 /// A number with provenance. `lastSyncedAt` is set when populated from an
 /// external source. Mirrors `SourcedNumber` in `src/state/schema.ts`.
 public struct SourcedNumber: Codable, Equatable, Sendable {
-    public var value: Double
+    public var value: Decimal
     public var source: Source
     public var lastSyncedAt: Date?
 
-    public init(value: Double, source: Source = .manual, lastSyncedAt: Date? = nil) {
+    public init(value: Decimal, source: Source = .manual, lastSyncedAt: Date? = nil) {
         self.value = value
         self.source = source
         self.lastSyncedAt = lastSyncedAt
     }
 
     /// Convenience for the common `{ value, source: .manual }` literal.
-    public static func manual(_ value: Double) -> SourcedNumber {
+    public static func manual(_ value: Decimal) -> SourcedNumber {
         SourcedNumber(value: value, source: .manual)
     }
 }
@@ -24,17 +40,17 @@ public struct SourcedNumber: Codable, Equatable, Sendable {
 public struct Debt: Codable, Equatable, Sendable, Identifiable {
     public var id: String
     public var name: String
-    public var balance: Double
-    public var apr: Double
-    public var minPayment: Double
+    public var balance: Decimal
+    public var apr: Decimal
+    public var minPayment: Decimal
     public var paid: Bool
 
     public init(
         id: String = Self.newID(),
         name: String = "",
-        balance: Double = 0,
-        apr: Double = 0,
-        minPayment: Double = 0,
+        balance: Decimal = 0,
+        apr: Decimal = 0,
+        minPayment: Decimal = 0,
         paid: Bool = false
     ) {
         self.id = id
@@ -53,16 +69,16 @@ public struct Debt: Codable, Equatable, Sendable, Identifiable {
 public struct Goal: Codable, Equatable, Sendable, Identifiable {
     public var id: String
     public var name: String
-    public var target: Double
-    public var saved: Double
-    public var horizonYears: Double
+    public var target: Decimal
+    public var saved: Decimal
+    public var horizonYears: Decimal
 
     public init(
         id: String = ShortID.make(),
         name: String = "",
-        target: Double = 0,
-        saved: Double = 0,
-        horizonYears: Double = 1
+        target: Decimal = 0,
+        saved: Decimal = 0,
+        horizonYears: Decimal = 1
     ) {
         self.id = id
         self.name = name
@@ -128,9 +144,9 @@ public enum IRAType: String, Codable, Sendable {
 public struct IRAData: Codable, Equatable, Sendable {
     public var type: IRAType
     public var ytdContribution: SourcedNumber
-    public var annualLimit: Double
+    public var annualLimit: Decimal
 
-    public init(type: IRAType = .roth, ytdContribution: SourcedNumber = .manual(0), annualLimit: Double) {
+    public init(type: IRAType = .roth, ytdContribution: SourcedNumber = .manual(0), annualLimit: Decimal) {
         self.type = type
         self.ytdContribution = ytdContribution
         self.annualLimit = annualLimit
@@ -153,9 +169,9 @@ public enum HSACoverage: String, Codable, Sendable {
 public struct HSAData: Codable, Equatable, Sendable {
     public var coverage: HSACoverage
     public var ytdContribution: SourcedNumber
-    public var annualLimit: Double
+    public var annualLimit: Decimal
 
-    public init(coverage: HSACoverage = .`self`, ytdContribution: SourcedNumber = .manual(0), annualLimit: Double) {
+    public init(coverage: HSACoverage = .`self`, ytdContribution: SourcedNumber = .manual(0), annualLimit: Decimal) {
         self.coverage = coverage
         self.ytdContribution = ytdContribution
         self.annualLimit = annualLimit
@@ -165,11 +181,11 @@ public struct HSAData: Codable, Equatable, Sendable {
 /// Payload for the SavePurchase node. Mirrors the `SavePurchase` shape in `NodeDataMap`.
 public struct SavePurchaseData: Codable, Equatable, Sendable {
     public var goalName: String
-    public var target: Double
+    public var target: Decimal
     public var saved: SourcedNumber
     public var byDate: String?
 
-    public init(goalName: String = "", target: Double = 0, saved: SourcedNumber = .manual(0), byDate: String? = nil) {
+    public init(goalName: String = "", target: Decimal = 0, saved: SourcedNumber = .manual(0), byDate: String? = nil) {
         self.goalName = goalName
         self.target = target
         self.saved = saved
@@ -179,11 +195,11 @@ public struct SavePurchaseData: Codable, Equatable, Sendable {
 
 /// Payload for the College (529) node. Mirrors the `College` shape in `NodeDataMap`.
 public struct CollegeData: Codable, Equatable, Sendable {
-    public var monthlyContribution: Double
+    public var monthlyContribution: Decimal
     public var balance: SourcedNumber
     public var targetAge: Int?
 
-    public init(monthlyContribution: Double = 0, balance: SourcedNumber = .manual(0), targetAge: Int? = nil) {
+    public init(monthlyContribution: Decimal = 0, balance: SourcedNumber = .manual(0), targetAge: Int? = nil) {
         self.monthlyContribution = monthlyContribution
         self.balance = balance
         self.targetAge = targetAge
